refactor: 変数を関数スコープにしまう
