+++ src/notify.ts
@@ -1,8 +1,9 @@
 import { isNew } from './lib/dateUtils';
 
+const modifiedAt = '2022-02-02';
+const newDays = 15;
+  
 export function notifyUpdate(): void {
-  const modifiedAt = '2022-02-02';
-  const newDays = 15;
   const aboutElement = document.getElementById('about');
 
   if (!isNew(new Date(modifiedAt), newDays)) return;
